+++ src/components/search/SearchInput/index.ts
@@ -1,0 +1,1 @@
+export * from "./SearchInput";
