+++ src/components/search/SearchItem/index.ts
@@ -1,0 +1,1 @@
+export * from "./SearchItem";
