+++ src/components/search/SearchPopular/SearchPopularItem.tsx
@@ -1,17 +1,17 @@
 interface Props {
   name: string;
-  tagId: number;
+  index: number;
 }
 
-export const SearchPopularItem = ({ name, tagId }: Props) => {
+export const SearchPopularItem = ({ name, index }: Props) => {
   return (
     <>
-      <li className="flex h-50 flex-row items-center justify-items-center">
+      <div className="flex h-50 flex-row items-center justify-items-center">
         <div className="mr-10 flex h-16 w-16 justify-center rounded-full border border-black bg-black text-white">
-          {tagId}
+          {index}
         </div>
         <div className="text-tag">{name}</div>
-      </li>
+      </div>
     </>
   );
 };
