+++ src/components/search/SearchRecent/index.ts
@@ -1,0 +1,1 @@
+export * from "./SearchRecent";
