+++ src/components/search/SearchResult/index.ts
@@ -1,0 +1,1 @@
+export * from "./SearchResult";
